+++ cursus_app/__init__.py
@@ -1,9 +1,10 @@
 from flask import Flask
 from flask_login import LoginManager
 
+from cursus_app.auth.views import auth_blueprint
+from cursus_app.course.views import course_blueprint
 from cursus_app.db import db
 from cursus_app.home.views import home_blueprint
-from cursus_app.auth.views import auth_blueprint
 from cursus_app.user.models import User
 
 
@@ -20,6 +21,7 @@ def create_app():
     login_manager.login_view = "auth.views"
     app.register_blueprint(home_blueprint)
     app.register_blueprint(auth_blueprint)
+    app.register_blueprint(course_blueprint)
 
     @login_manager.user_loader
     def load_user(user_id):
