+++ cursus_app/templates/home/index.html
@@ -25,7 +25,7 @@
             <div class="card-img-overlay">
                 <div class="card-body">
                     <h1 class="text-light" style="font-size: 64px;">Browse courses</h1>
-                    <a href="/courses" class="stretched-link"></a>
+                    <a href="{{url_for('course.index')}}" class="stretched-link"></a>
                 </div>
             </div>
           </div>
