--- cursus_app/topic/models.py
@@ -1,13 +1,0 @@
-from cursus_app.db import db
-
-
-class Topic(db.Model):
-    id = db.Column(
-        db.Integer, primary_key=True
-        )
-    name = db.Column(
-        db.String(32), unique=True, index=True, nullable=False
-        )
-
-    def __repr__(self):
-        return f"<Topic {self.name}>"
